Updated code for QR code for L2 approver
