+++ src/screens/approval/VerifyDetails.js
@@ -1,10 +1,15 @@
-import { Image, Platform, SafeAreaView, ScrollView, StyleSheet, Text, TouchableOpacity, View, ActivityIndicator, Alert } from 'react-native'
-import React, { useContext, useEffect, useState } from 'react'
+import { Image, Platform, SafeAreaView, ScrollView, StyleSheet, Text, TouchableOpacity, View, ActivityIndicator, Alert, ImageBackground, Dimensions } from 'react-native'
+import React, { useContext, useEffect, useState, useRef } from 'react'
 import { BASE_APP_URL, APP_LINK_NAME, APP_OWNER_NAME } from "@env"
 import UserContext from '../../../context/UserContext';
 import { encode } from 'base64-arraybuffer';
 import RNFS from 'react-native-fs';
 import Share from 'react-native-share';
+import LinearGradient from 'react-native-linear-gradient';
+import QRCode from 'react-native-qrcode-svg';
+import { captureRef } from 'react-native-view-shot';
+import 'react-native-get-random-values';
+import { BackgroundImage } from 'react-native-elements/dist/config';
 
 
 export const updateRecord = async (reportName, modified_data, token, id) => {
@@ -32,6 +37,16 @@ export const updateRecord = async (reportName, modified_data, token, id) => {
 
 
 
+
+
+const {height} = Dimensions.get('window')
+
+
+
+
+
+
+
 const VerifyDetails = ({ navigation, route }) => {
 
   const { user } = route.params;
@@ -41,6 +56,156 @@ const VerifyDetails = ({ navigation, route }) => {
   setEditData(user);
   const [loading, setLoading] = useState(true);
   const url = `${BASE_APP_URL}/${APP_OWNER_NAME}/${APP_LINK_NAME}/report/Approval_to_Visitor_Report/${user.ID}/Photo/download`
+
+  console.log("Screen Height:", height);
+    
+  let heightStyles;
+  if(height > 900){
+      heightStyles = normalScreen;
+  } else if(height>750){  heightStyles = mediumScreen;}
+  else{ heightStyles = smallScreen;}
+  const viewRef = useRef();
+  const [code, setCode] = useState('');
+  const codeGenrator = () => {
+      const newCode =  Math.floor(100000 + Math.random() * (999999 - 100001 + 1)).toString();
+      setCode(newCode);
+  };
+
+
+
+const ScreenshotQR = async () => {
+try{
+console.log('capturing view.......')
+const uri = await captureRef(viewRef, {
+format:'png',
+quality:0.8,
+
+});
+
+console.log('view captured Uri:', uri);
+
+// if (!uri){throw new Error('failed to capture, uri is undefined or null');
+// } 
+
+let base64Data = '';
+if (uri.startsWith('data:image/png;base64,')) {
+base64Data = uri.split('data:image/png;base64,')[1];
+} else if (uri.startsWith('file://')) {
+base64Data = await RNFS.readFile(uri, 'base64');
+} else {
+throw new Error(`Unexpected URI format: ${uri}`);
+}
+
+
+
+
+
+
+console.log('extracted base 64 data:', base64Data.length);
+
+
+if (!base64Data){
+  throw new Error('failed to extract base64 Data from URI');
+} 
+
+const postData = new FormData ();
+postData.append( 'file',{
+
+  uri: `data:image/png;base64, ${base64Data}`,
+  name: 'qrcode.png',
+  type: 'image/png',
+
+});
+
+const payload =  {
+  data: {
+      Generated_Passcode: code
+  }
+}
+
+const url1 = `${BASE_APP_URL}/${APP_OWNER_NAME}/${APP_LINK_NAME}/report/Approval_to_Visitor_Report/${user.ID}`
+console.log(url1);
+const response1 = await fetch(url1,{
+  method:'PATCH',
+  body: JSON.stringify(payload),
+  headers:{
+Authorization: `Zoho-oauthtoken ${accessToken}`,
+'Content-Type': 'application/json',
+
+
+
+  },
+}, console.log('posting to zoho....'));
+if (response1.ok) {
+  console.log('code posted successfully to Zoho.');
+  console.log("response",response1);
+} else {
+  console.log('Failed to post code to Zoho:', response1.status, response1.statusText, response1.ok);
+
+};
+
+
+
+
+const url = `${BASE_APP_URL}/${APP_OWNER_NAME}/${APP_LINK_NAME}/report/Approval_to_Visitor_Report/${user.ID}/Generated_QR_Code/upload`
+console.log(url);
+const response = await fetch(url,{
+  method:'POST',
+  body: postData,
+  headers:{
+Authorization: `Zoho-oauthtoken ${accessToken}`,
+'Content-Type': 'multipart/form-data',
+
+
+  },
+}, console.log('posting to zoho....'));
+
+if (response.ok) {
+  console.log('Image uploaded successfully to Zoho.');
+} else {
+  console.log('Failed to upload image to Zoho:', response.status, response.statusText);
+}
+} catch (error) {
+console.error('Error capturing and uploading QR code:', error);
+}
+};
+
+
+
+
+
+
+
+// const path = `${RNFS.DocumentDirectoryPath}/QRCode.png`;
+// console.log('Saving image to path:', path);form
+// await RNFS.writeFile(path, base64Data, 'base64');
+// console.log('image saved');
+
+
+// await Share.open({
+// URL: `file://${path}`,
+// type: 'image/png',
+
+// });
+// console.log('image shared');
+
+
+
+
+
+
+
+  useEffect(()=>{
+      codeGenrator();
+  }, []);
+  
+//  useEffect(()=>{
+// if(DummyImageScreen){
+//     ScreenshotQR()
+// }
+//  }, [DummyImageScreen]);
+
+
 
   const getImage = async () => {
     try {
@@ -86,6 +251,9 @@ const VerifyDetails = ({ navigation, route }) => {
         Referrer_Approval: "APPROVED",
         L2_Approval_Status: "APPROVED"
       }
+       
+     ScreenshotQR();
+      
     } else {
       updateField = {
         Referrer_Approval: "APPROVED"
@@ -171,53 +339,52 @@ const VerifyDetails = ({ navigation, route }) => {
     } catch (error) {
       Alert.alert('Error', error.message);
     }
-  };
+
+};
 
 
 
   console.log("User in verify details : ", user)
   return (
-    <SafeAreaView style={{ flex: 1, backgroundColor: "#FFF" }}>
+    <><SafeAreaView style={{ flex: 1, backgroundColor: "#FFF", zIndex:1 }}>
       {/* <View style={styles.header}>
-        <View style={styles.headerContainer}>
-          <Text style={styles.headertxt}>Visitor details</Text>
-          <TouchableOpacity onPress={() => navigation.navigate("EditVerifydetails", { user: user })} style={{ width: 40, height: 40, padding: 10, marginEnd: 20 }}>
-            <Image
-              source={edit}
-              style={{ width: 40, height: 40, }}
-            />
-          </TouchableOpacity>
-        </View>
-      </View> */}
+      <View style={styles.headerContainer}>
+        <Text style={styles.headertxt}>Visitor details</Text>
+        <TouchableOpacity onPress={() => navigation.navigate("EditVerifydetails", { user: user })} style={{ width: 40, height: 40, padding: 10, marginEnd: 20 }}>
+          <Image
+            source={edit}
+            style={{ width: 40, height: 40, }}
+          />
+        </TouchableOpacity>
+      </View>
+    </View> */}
       <ScrollView style={styles.scrollview}>
-        {
-          user?.Referrer_Approval === "PENDING APPROVAL" ? (
-            <View style={[styles.container, { marginTop: 20 }]}>
-              <View style={[styles.left, { width: "50%" }]}>
-                <TouchableOpacity style={styles.btnAccept} onPress={onApprove}>
-                  <Text style={styles.btntxt}>Approve</Text>
-                </TouchableOpacity>
-              </View>
-              <View style={styles.right}>
-                <TouchableOpacity style={styles.btnReject} onPress={onReject}>
-                  <Text style={styles.btntxt}>Reject</Text>
-                </TouchableOpacity>
-              </View>
-            </View>
-          ) : user?.Referrer_Approval === "APPROVED" ? (
-            <View style={{ width: "100%", padding: 10, marginLeft: "30%" }}>
-              <TouchableOpacity style={[styles.btnReject]} onPress={onReject}>
-                <Text style={[styles.btntxt]}>Reject</Text>
-              </TouchableOpacity>
-            </View>
-          ) : user?.Referrer_Approval === "DENIED" ? (
-            <View style={{ width: "100%", padding: 10, marginLeft: "15%" }}>
+        {user?.Referrer_Approval === "PENDING APPROVAL" ? (
+          <View style={[styles.container, { marginTop: 20 }]}>
+            <View style={[styles.left, { width: "50%" }]}>
               <TouchableOpacity style={styles.btnAccept} onPress={onApprove}>
                 <Text style={styles.btntxt}>Approve</Text>
               </TouchableOpacity>
             </View>
-          ) : null
-        }
+            <View style={styles.right}>
+              <TouchableOpacity style={styles.btnReject} onPress={onReject}>
+                <Text style={styles.btntxt}>Reject</Text>
+              </TouchableOpacity>
+            </View>
+          </View>
+        ) : user?.Referrer_Approval === "APPROVED" ? (
+          <View style={{ width: "100%", padding: 10, marginLeft: "30%" }}>
+            <TouchableOpacity style={[styles.btnReject]} onPress={onReject}>
+              <Text style={[styles.btntxt]}>Reject</Text>
+            </TouchableOpacity>
+          </View>
+        ) : user?.Referrer_Approval === "DENIED" ? (
+          <View style={{ width: "100%", padding: 10, marginLeft: "15%" }}>
+            <TouchableOpacity style={styles.btnAccept} onPress={onApprove}>
+              <Text style={styles.btntxt}>Approve</Text>
+            </TouchableOpacity>
+          </View>
+        ) : null}
 
         <View style={[styles.container, { marginTop: 20 }]}>
           <View style={styles.left}>
@@ -366,36 +533,701 @@ const VerifyDetails = ({ navigation, route }) => {
             <Text style={styles.value}>{user.Home_or_Office}</Text>
           </View>
         </View>
-        {
-          user.Referrer_Approval === "APPROVED" && user.L2_Approval_Status == "APPROVED" && user.Referrer_App_User_lookup.ID == loggedUser.userId ? (
-            <View style={[styles.container, { marginTop: 20, marginBottom: 20 }]}>
-              <View style={styles.left}>
-                <Text style={styles.label}>Generated QR Code</Text>
-              </View>
-              <View style={[styles.right, { justifyContent: "center", alignItems: "center", marginBottom: 20 }]}>
-                <Image source={{ uri: user.Generated_QR_Code }} style={{ height: 200, width: 200 }} />
-                <TouchableOpacity
-                  style={[
-                    styles.HomeButton,
-                    { backgroundColor: 'green' },
-                  ]}
-                  onPress={() => {
-                    onShare();
-                  }}>
-                  <Text style={[styles.wewe, styles.wewe1]}>Share</Text>
-                </TouchableOpacity>
-              </View>
+        {user.Referrer_Approval === "APPROVED" && user.L2_Approval_Status == "APPROVED" && user.Referrer_App_User_lookup.ID == loggedUser.userId ? (
+          <View style={[styles.container, { marginTop: 20, marginBottom: 20 }]}>
+            <View style={styles.left}>
+              <Text style={styles.label}>Generated QR Code</Text>
             </View>
+            <View style={[styles.right, { justifyContent: "center", alignItems: "center", marginBottom: 20 }]}>
+              <Image source={{ uri: user.Generated_QR_Code }} style={{ height: 200, width: 200 }} />
+              <TouchableOpacity
+                style={[
+                  styles.HomeButton,
+                  { backgroundColor: 'green' },
+                ]}
+                onPress={() => {
+                  onShare();
+                } }>
+                <Text style={[styles.wewe, styles.wewe1]}>Share</Text>
+              </TouchableOpacity>
+            </View>
+          </View>
 
-          ) : null
-        }
+        ) : null}
 
       </ScrollView>
-    </SafeAreaView>
+    </SafeAreaView><View style={[heightStyles.hidden]}>
+        <View ref={viewRef} style={[heightStyles.container]}><View style={{ flex: 1 }}>
+          <View style={[heightStyles.qrCodeContainer]}>
+            <Text style={[heightStyles.title]}>L1 approver has invited you</Text>
+            <Text style={[heightStyles.text]}>Show this QR code or OTP to the guard at the gate</Text>
+            {code ? (
+              <QRCode value={code} size={160} />
+            ) : (<Text>Genrating Qr code....</Text>)}
+            <Text style={[heightStyles.middleText]}>---OR---</Text>
+            <View style={[heightStyles.codeBackdrop]}>
+              <Text style={[heightStyles.code]}>{code}</Text>
+              <View style={[heightStyles.BottomtextContainer]}>
+                <Text style={[heightStyles.dateOfArrivalText]}>{user.Date_of_Visit}</Text>
+                <Text style={[heightStyles.Bottomtext]}>Sri Sathya Sai Grama -</Text>
+                <Text style={[heightStyles.Bottomtext]}>Muddenahalli Rd,</Text>
+                <Text style={[heightStyles.Bottomtext]}> Karnataka 562103,</Text>
+                <View style={{ flex: 1 }}>
+
+                </View>
+              </View>
+            </View>
+            <View style={{ flex: 0.7 }}><ImageBackground style={[heightStyles.BottomImage]} source={require('../../../src/assets/ashramQrScreen.jpg')}>
+              <LinearGradient colors={['rgba(255,255,255,1)', 'rgba(255,255,255,0)']} style={[heightStyles.gradient, heightStyles.topGradient]} /><LinearGradient colors={['rgba(255,255,255,0)', 'rgba(255,255,255,1)']} style={[heightStyles.gradient, heightStyles.bottomGradient]} /></ImageBackground>
+
+              <ImageBackground style={[heightStyles.BottomLogoImage]} source={require('../../../src/assets/SSG_OWOF.png')}></ImageBackground>
+
+
+            </View>
+          </View>
+
+          {/* <View style={[styles.Buttons]}>
+<Button title="Generate new Qr code"  onPress={codeGenrator}/>
+<Button title="Upload Qr code"  onPress={ScreenshotQR}></Button>
+
+</View> */}
+        </View>
+        </View>
+      </View></>
   )
 }
 
 export default VerifyDetails
+
+const mediumScreen = StyleSheet.create({
+  hidden:{
+ opacity:0,
+ position:'absolute',
+ zIndex:0,
+
+  },
+
+  gradient:{
+    ...StyleSheet.absoluteFillObject,
+    position:'absolute',
+    left:0,
+    top:0,
+},
+
+
+  topGradient:{
+    top:0,
+    height:'180%'
+            },
+          
+    bottomGradient:{
+          bottom:0,
+          height:'9%',
+          backgroundColor:'#F9ECDF'
+            },
+
+
+
+  BottomImage:{
+
+  flex: 1,
+  position: 'relative',
+  justifyContent: 'flex-end',
+  alignSelf:'center',
+  width: 385,
+  height: 120, // height as a percentage of screen height
+  position: 'absolute',
+  bottom: -76,
+  
+  },
+
+
+  BottomLogoImage:{
+
+    flex: 1,
+    position: 'relative',
+    justifyContent: 'flex-end',
+    alignSelf:'center',
+    width: 145,
+    height: 95, // height as a percentage of screen height
+    position: 'absolute',
+    bottom: -76,
+    
+    },
+  
+  pageContainer:{
+      backgroundColor:'white'
+  },
+  
+   container: {
+      flex:1,
+      justifyContent:'center',
+      alignItems:'center',
+      backgroundColor:'#F9ECDF',
+      width:385,
+      height:612
+  
+  
+    
+  
+  },
+  
+  title:{
+      fontSize:30,
+      textAlign: 'center',
+      margin:10,
+      color:'#6E260E',
+      fontWeight:'bold'
+  },
+  
+  code:{
+      fontSize:35,
+      textAlign:'center',
+      color:'brown',
+     
+  },
+  
+  codeBackdrop:{
+      marginTop:12,
+      backgroundColor:'pink',
+      borderRadius:20,
+      flexGrow:0,
+      width:170,
+      height:50
+      
+  
+  },
+  
+  
+  text:{
+      fontSize:14,
+      textAlign:'center',
+      color:'#6E260E',
+      marginBottom:10
+     
+  },
+  
+  middleText:{
+  fontSize:17,
+  color:'#6E260E',
+  marginTop:10
+  
+  
+  },
+  
+  BottomtextContainer:{
+  marginTop:15
+  
+  
+  },
+  
+  
+  Bottomtext:{
+      fontSize:10,
+      textAlign:'center',
+      color:'#6E260E',
+      
+      },
+  
+  
+  dateOfArrivalText:{
+      color:'#6E260E',
+      fontWeight:'bold',
+     alignSelf:'center',
+     fontSize:20,
+     
+  
+  },
+  
+  qrCodeContainer:{
+  flex:0.92,
+      alignItems:'center',
+      justifyContent:'center',
+      
+  },
+  
+  Buttons:{
+      marginTop:100
+  },
+})
+
+
+const smallScreen = StyleSheet.create({
+  hidden:{
+      opacity:0,
+      position:'absolute',
+      zIndex:0,
+   
+   
+       },
+
+topGradient:{
+  top:0,
+  height:'180%'
+          },
+        
+  bottomGradient:{
+        bottom:0,
+        height:'9%',
+        backgroundColor:'#F9ECDF'
+          },
+  
+  BottomImage:{
+
+  flex: 1,
+  position: 'relative',
+  justifyContent: 'flex-end',
+  alignSelf:'center',
+  width: 440,
+  height: 90, // height as a percentage of screen height
+  position: 'absolute',
+  bottom: -35
+  
+  
+  },
+
+
+  BottomLogoImage:{
+
+    
+    },
+  
+
+  gradient:{
+      ...StyleSheet.absoluteFillObject,
+      ...StyleSheet.absoluteFillObject,
+      position:'absolute',
+      left:0,
+      top:0,
+  },
+  
+  pageContainer:{
+      backgroundColor:'white'
+  },
+  
+   container: {
+      flex:0,
+      justifyContent:'center',
+      alignItems:'center',
+      backgroundColor:'#F9ECDF',
+      width:430,
+      height:570
+  
+    
+  
+  },
+  
+  title:{
+      fontSize:30,
+      textAlign: 'center',
+      margin:10,
+      color:'#6E260E',
+      fontWeight:'bold'
+  },
+  
+  code:{
+      fontSize:35,
+      textAlign:'center',
+      color:'brown',
+     
+  },
+  
+  codeBackdrop:{
+      marginTop:12,
+      backgroundColor:'pink',
+      borderRadius:20,
+      flexGrow:0,
+      width:170,
+      height:50
+      
+  
+  },
+  
+  
+  text:{
+      fontSize:14,
+      textAlign:'center',
+      color:'#6E260E',
+      marginBottom:10
+     
+  },
+  
+  middleText:{
+  fontSize:17,
+  color:'#6E260E',
+  marginTop:10
+  
+  
+  },
+  
+  BottomtextContainer:{
+  marginTop:15
+  
+  
+  },
+  
+  
+  Bottomtext:{
+      fontSize:10,
+      textAlign:'center',
+      color:'#6E260E',
+      
+      },
+  
+  
+  dateOfArrivalText:{
+      color:'#6E260E',
+      fontWeight:'bold',
+     alignSelf:'center',
+     fontSize:20,
+     
+  
+  },
+  
+  qrCodeContainer:{
+  flex:1,
+      alignItems:'center',
+      justifyContent:'center',
+      
+  },
+  
+  Buttons:{
+      marginTop:100
+  },
+})
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+const normalScreen = StyleSheet.create({
+
+  hidden:{
+      opacity:0,
+      position:'absolute',
+      zIndex:0,
+ 
+   
+       },
+
+
+
+
+  gradient:{
+    ...StyleSheet.absoluteFillObject,
+      position:'absolute',
+      left:0,
+      top:0,
+  },
+
+
+  topGradient:{
+top:0,
+height:'180%'
+  },
+
+  bottomGradient:{
+bottom:0,
+height:'10%',
+backgroundColor:'#F9ECDF'
+
+
+  },
+
+
+
+pageContainer:{
+  backgroundColor:'white'
+},
+
+BottomLogoImage:{
+
+  flex: 1,
+  position: 'relative',
+  justifyContent: 'flex-end',
+  alignSelf:'center',
+  width: 170,
+  height: 120, // height as a percentage of screen height
+  position: 'absolute',
+  bottom: -40,
+alignItems:'center'
+  },
+
+
+container: {
+  flex:1,
+  justifyContent:'center',
+  alignItems:'center',
+  backgroundColor:'#F9ECDF',
+  width:450,
+  height:780,
+ 
+
+
+},
+
+title:{
+  fontSize:30,
+  textAlign: 'center',
+  margin:10,
+  color:'#6E260E',
+  fontWeight:'bold'
+},
+
+code:{
+  fontSize:35,
+  textAlign:'center',
+  color:'brown',
+ 
+},
+
+codeBackdrop:{
+  marginTop:12,
+  backgroundColor:'pink',
+  borderRadius:20,
+  flexGrow:0,
+  width:170,
+  height:50
+  
+
+},
+
+
+text:{
+  fontSize:14,
+  textAlign:'center',
+  color:'#6E260E',
+  marginBottom:10
+ 
+},
+
+middleText:{
+fontSize:17,
+color:'#6E260E',
+marginTop:10
+
+
+},
+
+BottomtextContainer:{
+marginTop:19
+
+
+},
+
+
+Bottomtext:{
+  fontSize:18,
+  textAlign:'center',
+  color:'#6E260E',
+  
+  },
+
+
+dateOfArrivalText:{
+  color:'#6E260E',
+  fontWeight:'bold',
+ alignSelf:'center',
+ fontSize:20,
+ 
+
+},
+
+qrCodeContainer:{
+flex:1,
+  alignItems:'center',
+  justifyContent:'center',
+  
+},
+
+Buttons:{
+  marginTop:100
+},
+
+BottomImage:{
+
+  flex: 1,
+  position: 'relative',
+  justifyContent: 'flex-end',
+  alignSelf:'center',
+  width: 500,
+  height: 200, // height as a percentage of screen height
+  position: 'absolute',
+  bottom: -78,
+  
+
+
+
+}
+
+});
+
+
+
+
+// const mainStyles = StyleSheet.create({
+
+//     pageContainer:{
+//         backgroundColor:'white'
+//     },
+  
+//      container: {
+//         flex:1,
+//         justifyContent:'center',
+//         alignItems:'center',
+//         backgroundColor:'#F9ECDF',
+  
+  
+    
+  
+//     },
+  
+//     title:{
+//         fontSize:30,
+//         textAlign: 'center',
+//         margin:10,
+//         color:'#6E260E',
+//         fontWeight:'bold'
+//     },
+  
+//     code:{
+//         fontSize:35,
+//         textAlign:'center',
+//         color:'brown',
+     
+//     },
+  
+//     codeBackdrop:{
+//         marginTop:12,
+//         backgroundColor:'pink',
+//         borderRadius:20,
+//         flexGrow:0,
+//         width:170,
+//         height:50
+      
+  
+//     },
+  
+  
+//     text:{
+//         fontSize:14,
+//         textAlign:'center',
+//         color:'#6E260E',
+//         marginBottom:10
+     
+//     },
+  
+//     middleText:{
+//     fontSize:17,
+//     color:'#6E260E',
+//     marginTop:10
+  
+  
+//     },
+  
+//     BottomtextContainer:{
+//     marginTop:15
+  
+  
+//     },
+  
+  
+//     Bottomtext:{
+//         fontSize:10,
+//         textAlign:'center',
+//         color:'#6E260E',
+      
+//         },
+  
+  
+//     dateOfArrivalText:{
+//         color:'#6E260E',
+//         fontWeight:'bold',
+//        alignSelf:'center',
+//        fontSize:20,
+     
+  
+//     },
+  
+//     qrCodeContainer:{
+//     flex:1,
+//         alignItems:'center',
+//         justifyContent:'center',
+      
+//     },
+  
+//     Buttons:{
+//         marginTop:100
+//     },
+  
+//     BottomImage:{
+  
+//         flex: 1,
+//         position: 'relative',
+//         justifyContent: 'flex-end',
+//         alignSelf:'center',
+//         width: 500,
+//         height: 200, // height as a percentage of screen height
+//         position: 'absolute',
+//         bottom: -198,
+      
+  
+  
+  
+//     }
+  
+//     });
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 
 const styles = StyleSheet.create({
   header: {
